educa/requirements.txt
Django==2.2.3
pytz==2018.5
dj-config-url==0.1.1
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-mathfilters==0.4.0
gunicorn==19.9.0
psycopg2==2.7.5
whitenoise==4.1
djangorestframework==3.9.1
Pillow==5.3.0
psycopg2-binary==2.7.6.1
django-braces==1.13.0
django-embed-video==1.1.2
python-memcached==1.59
django-memcache-status==1.3
social-auth-app-django==2.1.0
sqlparse==0.3.0

celery==4.2.1
amqp==2.3.2
billiard==3.5.0.4
celery==4.2.1
kombu==4.2.1
vine==1.1.4
idna==2.7

braintree==3.55.0

certifi==2018.10.15
chardet==3.0.4

requests==2.20.0
urllib3==1.24

ibm-watson==3.2.0
django-twilio
South==1.0.2
phonenumbers==8.10.15

django-embed-video==1.1.2
python-memcached==1.59
django-memcache-status==1.3
social-auth-app-django==2.1.0
django-storages==1.7.1
boto3
